更新所有相关Nuget包到最新版本 更新数据包到最新数据 新增ListHelper帮助类将列表转换为树形结构 新增用户删除后真删用户角色数据 新增角色删除后真删角色权限数据 新增角色唯一编码 调整data.json数据主键为雪花漂移Id，扁平化数据调整为树形结构数据。 修复个人信息上传头像失败无错误提示的问题 优化角色权限界面保存逻辑更改为允许空权限保存，刷新权限禁止保存。 优化角色权限刷新角色列表时同时刷权限列表

+++ Admin.Core.Services/Admin/Permission/Input/PermissionAddApiInput.cs
@@ -12,12 +12,12 @@ namespace Admin.Core.Service.Admin.Permission.Input
         /// <summary>
         /// 父级节点
         /// </summary>
-        public int ParentId { get; set; }
+        public long ParentId { get; set; }
 
         /// <summary>
         /// 接口
         /// </summary>
-        public int? ApiId { get; set; }
+        public long? ApiId { get; set; }
 
         /// <summary>
         /// 权限名称
