+++ Admin.Core.Services/Admin/Permission/Input/PermissionAddMenuInput.cs
@@ -12,7 +12,7 @@ namespace Admin.Core.Service.Admin.Permission.Input
         /// <summary>
         /// 父级节点
         /// </summary>
-        public int ParentId { get; set; }
+        public long ParentId { get; set; }
 
         /// <summary>
         /// 视图
