+++ Admin.Core.Services/Admin/Permission/Input/PermissionAssignInput.cs
@@ -6,9 +6,9 @@ namespace Admin.Core.Service.Admin.Permission.Input
     public class PermissionAssignInput
     {
         [Required(ErrorMessage = "角色不能为空！")]
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
 
         [Required(ErrorMessage = "权限不能为空！")]
-        public List<int> PermissionIds { get; set; }
+        public List<long> PermissionIds { get; set; }
     }
 }
