+++ Admin.Core.Services/Admin/Role/Input/RoleUpdateInput.cs
@@ -3,12 +3,12 @@
     /// <summary>
     /// 修改
     /// </summary>
-    public partial class RoleUpdateInput :RoleAddInput
+    public partial class RoleUpdateInput : RoleAddInput
     {
         /// <summary>
         /// 接口Id
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 版本
